+++ packages/query-tools/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @neo4j-cypher/query-tools
 
+## 2.0.0-next.36
+
+### Patch Changes
+
+- Updated dependencies [50f061e]
+  - @neo4j-cypher/language-support@2.0.0-next.36
+
 ## 2.0.0-next.35
 
 ### Patch Changes