+++ packages/react-codemirror/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @neo4j-cypher/react-codemirror
 
+## 2.0.0-next.39
+
+### Patch Changes
+
+- Updated dependencies [50f061e]
+  - @neo4j-cypher/language-support@2.0.0-next.36
+  - @neo4j-cypher/lint-worker@1.10.1-next.13
+
 ## 2.0.0-next.38
 
 ### Patch Changes